test(e2e): keep multiarch qualification linear
