+++ test/e2e/mock-parity.json
@@ -3,6 +3,15 @@
   "version": 1,
   "entries": [
     {
+      "live": "test/e2e/live/managed-image-multiarch-startup.test.ts",
+      "fast": [
+        "test/e2e/support/base-image-publication.test.ts",
+        "test/managed-image-publication-workflow.test.ts",
+        "test/pr-risk-plan.test.ts",
+        "test/protected-managed-image-contract.test.ts"
+      ]
+    },
+    {
       "live": "test/e2e/live/hermes-gpu-startup.test.ts",
       "fast": [
         "test/e2e/support/hermes-gpu-startup-fallback.test.ts",
